backend/requirements.txt
fastapi 
uvicorn 
Pillow 
google-generativeai 
redis 
prometheus-fastapi-instrumentator 
pytest 
httpx
python-multipart
python-dotenv